moved GetClientList
